// src/types/walletResultStatus.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type WalletResultStatus = typeof WalletResultStatus[keyof typeof WalletResultStatus];


export const WalletResultStatus = {
  ACTIVE: 'ACTIVE',
  LOCKED: 'LOCKED',
  FROZEN: 'FROZEN',
} as const;